+++ src/localllm/models.py
@@ -16,9 +16,13 @@ def localllm_download_model(type: str = "gemma-3-270m-it-Q8_0", model_dir: Optio
         - "gemma-3-270m-it-Q8_0": Google Gemma 3 270M it model (Q8_0 quantization)
         - "gemma-3-270m-it-qat-Q4_0": Google Gemma 3 270M it model (Q4_0 quantization)
         - "gemma-3-1b-it-Q8_0": Google Gemma 3 1B it model (Q8_0 quantization)
-        - "gemma-3-4b-it-Q4_K_M": Google Gemma 3 4b it model (Q4_K_M quantization)        
+        - "gemma-3-4b-it-Q4_K_M": Google Gemma 3 4b it model (Q4_K_M quantization)     
+        - "gemma-3-12b-it-qat-Q4_0": Google Gemma 3 12B it model (Q4_0 quantization)
         - "GLM-4.6V-Flash-Q4_K_M": GLM 4.6V Flash model (Q4_K_M quantization)
         - "translategemma-4b-it-q8_0": TranslateGemma 4B it model (Q8_0 quantization)
+        - "LFM2.5-1.2B-Instruct-Q4_K_M": LFM2.5 1.2B Instruct model (Q4_K_M quantization)
+        - "Qwen3-8B-Q4_K_M": Qwen 3 8B model (Q4_K_M quantization)
+        
 
     model_dir : str or None, default=None
         Directory where the model should be stored. If None, uses the path set in environment variable LOCALLLM_MODEL_DIR
@@ -55,8 +59,11 @@ def localllm_download_model(type: str = "gemma-3-270m-it-Q8_0", model_dir: Optio
     >>> model_path = localllm_download_model("gemma-3-270m-it-qat-Q4_0", overwrite=True, trace = False)
     >>> model_path = localllm_download_model("gemma-3-1b-it-Q8_0", overwrite=True, trace = False)
     >>> model_path = localllm_download_model("gemma-3-4b-it-Q4_K_M", overwrite=True, trace = False)
+    >>> model_path = localllm_download_model("gemma-3-12b-it-qat-Q4_0", overwrite=True, trace = False)
     >>> model_path = localllm_download_model("GLM-4.6V-Flash-Q4_K_M", overwrite=True, trace = False)
     >>> model_path = localllm_download_model("translategemma-4b-it-q8_0", overwrite=True, trace = False)
+    >>> model_path = localllm_download_model("LFM2.5-1.2B-Instruct-Q4_K_M", overwrite=True, trace = False)
+    >>> model_path = localllm_download_model("Qwen3-8B-Q4_K_M", overwrite=True, trace = False)
 
     Notes
     -----
@@ -80,8 +87,13 @@ def localllm_download_model(type: str = "gemma-3-270m-it-Q8_0", model_dir: Optio
         },
         "gemma-3-4b-it-Q4_K_M": {
             "url": "https://huggingface.co/bartowski/google_gemma-3-4b-it-GGUF/resolve/main/google_gemma-3-4b-it-Q4_K_M.gguf",
+            #"url": "https://huggingface.co/unsloth/gemma-3-4b-it-GGUF/resolve/main/gemma-3-4b-it-Q4_K_M.gguf",
             "filename": "google_gemma-3-4b-it-Q4_K_M.gguf",
         },
+        "gemma-3-12b-it-qat-Q4_0": {
+            "url": "https://huggingface.co/bartowski/google_gemma-3-12b-it-qat-GGUF/resolve/main/google_gemma-3-12b-it-qat-Q4_0.gguf",            
+            "filename": "google_gemma-3-12b-it-qat-Q4_0.gguf",
+        },           
         "GLM-4.6V-Flash-Q4_K_M": {
             "url": "https://huggingface.co/ggml-org/GLM-4.6V-Flash-GGUF/resolve/main/GLM-4.6V-Flash-Q4_K_M.gguf",
             "filename": "GLM-4.6V-Flash-Q4_K_M.gguf",
@@ -90,6 +102,14 @@ def localllm_download_model(type: str = "gemma-3-270m-it-Q8_0", model_dir: Optio
             "url": "https://huggingface.co/NikolayKozloff/translategemma-4b-it-Q8_0-GGUF/resolve/main/translategemma-4b-it-q8_0.gguf",
             "filename": "translategemma-4b-it-q8_0.gguf",
         },
+        "LFM2.5-1.2B-Instruct-Q4_K_M": {
+            "url": "https://huggingface.co/LiquidAI/LFM2.5-1.2B-Instruct-GGUF/resolve/main/LFM2.5-1.2B-Instruct-Q4_K_M.gguf",
+            "filename": "LFM2.5-1.2B-Instruct-Q4_K_M.gguf",
+        },      
+        "Qwen3-8B-Q4_K_M": {
+            "url": "https://huggingface.co/bartowski/Qwen_Qwen3-8B-GGUF/resolve/main/Qwen_Qwen3-8B-Q4_K_M.gguf",            
+            "filename": "Qwen_Qwen3-8B-Q4_K_M.gguf",
+        },                      
     }
 
     # Validate model type